requirements.txt
colored
flask
flask-cors
jupyter
python-dotenv
python-dotenv
requests
